Refactor: Extract partial components and hooks in consistent folders
